Add support for Redis 6 and 7 (#40)
* WIP: Add support for Redis 6+ using built-in TLS

* WIP: Fix bats tests

* WIP: Update replication tests

* Update tests to work with previous versions

* Add TLS protocol tests for Redis 6

* Add nordb and aof images

* Replace aof and nordb test files with links

* Add Redis 6 to build matrix

* Add support for Redis 7

* Add aof and nordb images

* Increase replication test timeout

* Update replication test

* Clarify --restore support

* Update README and latest tag

+++ bin/run-database.sh
@@ -15,7 +15,7 @@ MASTER_FORWARD_PORT=8765
 MASTER_FORWARD_CONF="${DATA_DIRECTORY}/master-tunnel.conf"
 
 ensure_ssl_material() {
-  if [ -n "${SSL_CERTIFICATE:-}" ] && [ -n "${SSL_KEY:-}" ]; then
+  if [[ -n "${SSL_CERTIFICATE:-}" ]] && [[ -n "${SSL_KEY:-}" ]]; then
     # Nothing to do!
     return
   fi
@@ -96,10 +96,15 @@ start_redis_cli() {
       port="$SSL_PORT"
     fi
 
-    stunnel_dir="$(create_ephemeral_tunnel "$host" "$port")"
-    redis-cli -h "127.0.1" -p "$(cat "${stunnel_dir}/port")" -a "$password" "$@"
-    kill -TERM "$(cat "${stunnel_dir}/pid")"
-    rm -r "${stunnel_dir}"
+    if [[ -n "${INTEGRATED_TLS:-}" ]]; then
+      # shellcheck disable=SC2154
+      redis-cli -h "$host" -p "$port" -a "$password" --tls "$@"
+    else
+      stunnel_dir="$(create_ephemeral_tunnel "$host" "$port")"
+      redis-cli -h "127.0.1" -p "$(cat "${stunnel_dir}/port")" -a "$password" "$@"
+      kill -TERM "$(cat "${stunnel_dir}/pid")"
+      rm -r "${stunnel_dir}"
+    fi
 
   else
     echo "Unknown protocol: $protocol"
@@ -109,24 +114,33 @@ start_redis_cli() {
 start_server() {
   ensure_ssl_material
 
-  STUNNEL_DIRECTORY="$(mktemp -d -p "$STUNNEL_ROOT_DIRECTORY")"
-  export STUNNEL_DIRECTORY
+  if [[ -n "${INTEGRATED_TLS:-}" ]]; then
+    TLS_CERT_FILE="$(mktemp)"
+    TLS_KEY_FILE="$(mktemp)"
 
-  # Set up SSL using stunnel.
-  SSL_CERT_FILE="$(mktemp -p "$STUNNEL_DIRECTORY")"
-  echo "$SSL_CERTIFICATE" > "$SSL_CERT_FILE"
-  unset SSL_CERTIFICATE
+    echo "$SSL_CERTIFICATE" > "$TLS_CERT_FILE"
+    echo "$SSL_KEY" > "$TLS_KEY_FILE"
 
-  SSL_KEY_FILE="$(mktemp -p "$STUNNEL_DIRECTORY")"
-  echo "$SSL_KEY" > "$SSL_KEY_FILE"
-  unset SSL_KEY
+    chown "${REDIS_USER}:${REDIS_USER}" "$TLS_CERT_FILE" "$TLS_KEY_FILE"
+  else
+    STUNNEL_DIRECTORY="$(mktemp -d -p "$STUNNEL_ROOT_DIRECTORY")"
+    export STUNNEL_DIRECTORY
 
-  STUNNEL_TUNNELS_DIRECTORY="${STUNNEL_DIRECTORY}/tunnels"
-  mkdir "$STUNNEL_TUNNELS_DIRECTORY"
+    # Set up SSL using stunnel.
+    SSL_CERT_FILE="$(mktemp -p "$STUNNEL_DIRECTORY")"
+    echo "$SSL_CERTIFICATE" > "$SSL_CERT_FILE"
+    unset SSL_CERTIFICATE
 
-  REDIS_TUNNEL_FILE="${STUNNEL_TUNNELS_DIRECTORY}/redis.conf"
+    SSL_KEY_FILE="$(mktemp -p "$STUNNEL_DIRECTORY")"
+    echo "$SSL_KEY" > "$SSL_KEY_FILE"
+    unset SSL_KEY
 
-  cat > "$REDIS_TUNNEL_FILE" <<EOF
+    STUNNEL_TUNNELS_DIRECTORY="${STUNNEL_DIRECTORY}/tunnels"
+    mkdir "$STUNNEL_TUNNELS_DIRECTORY"
+
+    REDIS_TUNNEL_FILE="${STUNNEL_TUNNELS_DIRECTORY}/redis.conf"
+
+    cat > "$REDIS_TUNNEL_FILE" <<EOF
 [redis]
 accept = ${SSL_PORT}
 connect = ${REDIS_PORT}
@@ -134,8 +148,9 @@ cert = ${SSL_CERT_FILE}
 key = ${SSL_KEY_FILE}
 EOF
 
-  if [[ -f "$MASTER_FORWARD_CONF" ]]; then
-    cp "$MASTER_FORWARD_CONF" "${STUNNEL_TUNNELS_DIRECTORY}/master-tunnel.conf"
+    if [[ -f "$MASTER_FORWARD_CONF" ]]; then
+      cp "$MASTER_FORWARD_CONF" "${STUNNEL_TUNNELS_DIRECTORY}/master-tunnel.conf"
+    fi
   fi
 
   # Finally, we force-chown the data directory and its contents. There won't be many
@@ -145,12 +160,21 @@ EOF
 
   touch "$ARGUMENT_FILE" # don't crash and burn if initialize wasn't called.
 
-  if [ -n "${MAX_MEMORY:-}" ]; then
+  if [[ -n "${MAX_MEMORY:-}" ]]; then
     echo "--maxmemory-policy allkeys-lru" >> "$ARGUMENT_FILE"
     echo "--maxmemory ${MAX_MEMORY}" >> "$ARGUMENT_FILE"
   fi
 
-  exec supervisord -c "/etc/supervisord.conf"
+  if [[ -n "${INTEGRATED_TLS:-}" ]]; then
+    echo "--tls-port ${SSL_PORT}" >> "$ARGUMENT_FILE"
+    echo "--tls-cert-file ${TLS_CERT_FILE}" >> "$ARGUMENT_FILE"
+    echo "--tls-key-file ${TLS_KEY_FILE}" >> "$ARGUMENT_FILE"
+    echo "--tls-auth-clients no" >> "$ARGUMENT_FILE"
+
+    exec sudo -E -u "$REDIS_USER" redis-wrapper
+  else
+    exec supervisord -c "/etc/supervisord.conf"
+  fi
 }
 
 
@@ -161,6 +185,7 @@ elif [[ "$1" == "--initialize" ]]; then
   echo "--requirepass $PASSPHRASE" > "$ARGUMENT_FILE"
 
   touch "$CONFIG_EXTRA_FILE"
+
   if [[ -n "${REDIS_NORDB:-}" ]]; then
     echo 'appendonly no' >> "$CONFIG_EXTRA_FILE"
     echo 'save ""' >> "$CONFIG_EXTRA_FILE"
@@ -188,13 +213,21 @@ elif [[ "$1" == "--initialize-from" ]]; then
     fi
 
   elif [[ "$protocol" = "rediss://" ]]; then
-    create_tunnel_configuration redis-master \
-      "127.0.0.1" "$MASTER_FORWARD_PORT" \
-      "$host" "$port" \
-      > "$MASTER_FORWARD_CONF"
+    if [[ -n "${INTEGRATED_TLS:-}" ]]; then
+      if [[ -z "$port" ]]; then
+        port="$SSL_PORT"
+      fi
 
-    host="127.0.0.1"
-    port="$MASTER_FORWARD_PORT"
+      tls_replication='yes'
+    else
+      create_tunnel_configuration redis-master \
+        "127.0.0.1" "$MASTER_FORWARD_PORT" \
+        "$host" "$port" \
+        > "$MASTER_FORWARD_CONF"
+
+      host="127.0.0.1"
+      port="$MASTER_FORWARD_PORT"
+    fi
   else
     echo "Unknown protocol: $protocol"
   fi
@@ -203,6 +236,10 @@ elif [[ "$1" == "--initialize-from" ]]; then
     echo "--requirepass $password" > "$ARGUMENT_FILE"
     echo "--slaveof $host ${port}" >> "$ARGUMENT_FILE"
     echo "--masterauth $password" >> "$ARGUMENT_FILE"
+    if [[ -n "${tls_replication:-}" ]]; then
+      echo '--tls-replication yes' >> "$ARGUMENT_FILE"
+      echo "--tls-ca-cert-dir ${SSL_CERTS_DIRECTORY}" >> "$ARGUMENT_FILE"
+    fi
   }
 
 elif [[ "$1" == "--client" ]]; then
@@ -212,7 +249,11 @@ elif [[ "$1" == "--client" ]]; then
 
 elif [[ "$1" == "--dump" ]]; then
   [ -z "$2" ] && echo "docker run -i aptible/redis --dump redis://... > dump.rdb" && exit
-  start_redis_cli "$2"  --rdb "$DUMP_FILENAME"
+  shift
+
+  # Redis 6+ outputs additional data to stdout.
+  # Redirect it to stderr so that stdout only contains the dump
+  start_redis_cli "$@" --rdb "$DUMP_FILENAME" >&2
 
   #shellcheck disable=SC2015
   [ -e /dump-output ] && exec 3>/dump-output || exec 3>&1
@@ -221,11 +262,12 @@ elif [[ "$1" == "--dump" ]]; then
 
 elif [[ "$1" == "--restore" ]]; then
   [ -z "$2" ] && echo "docker run -i aptible/redis --restore redis://... < dump.rdb" && exit
+  shift
 
   #shellcheck disable=SC2015
   [ -e /restore-input ] && exec 3</restore-input || exec 3<&0
   cat > "$DUMP_FILENAME" <&3
-  rdb --command protocol "$DUMP_FILENAME" | start_redis_cli "$2" --pipe
+  rdb --command protocol "$DUMP_FILENAME" | start_redis_cli "$@" --pipe
   rm "$DUMP_FILENAME"
 
 elif [[ "$1" == "--readonly" ]]; then
